+++ spec/value_objects/conviction_type_spec.rb
@@ -455,7 +455,7 @@ RSpec.describe ConvictionType do
 
       it { expect(conviction_type.skip_length?).to eq(false) }
       it { expect(conviction_type.relevant_order?).to eq(true) }
-      it { expect(conviction_type.calculator_class).to eq(Calculators::Motoring::Youth::Disqualification) }
+      it { expect(conviction_type.calculator_class).to eq(Calculators::DisqualificationCalculator::Youths) }
     end
 
     context 'YOUTH_MOTORING_FINE' do
@@ -481,7 +481,7 @@ RSpec.describe ConvictionType do
 
       it { expect(conviction_type.skip_length?).to eq(false) }
       it { expect(conviction_type.relevant_order?).to eq(true) }
-      it { expect(conviction_type.calculator_class).to eq(Calculators::Motoring::Adult::Disqualification) }
+      it { expect(conviction_type.calculator_class).to eq(Calculators::DisqualificationCalculator::Adults) }
     end
 
     context 'ADULT_MOTORING_FINE' do
